+++ shared/store-chunk.ts
@@ -3,7 +3,7 @@ import { CacheMap } from "./cache";
 import { delay } from "./shared";
 import { Store, IStore } from "./store";
 
-function getGroupId(s: string) {
+export function getGroupId(s: string) {
   for(var i = 0, h = 0; i < s.length; i++)
     h = Math.imul(17, h) + s.charCodeAt(i) | 0;
   return (Math.abs(h) % 256).toString(16);
@@ -58,7 +58,6 @@ export class ChunkedStore extends IStore {
     } else {
       const group = (await this.store.get('g' + gid) || {}) as any;
       this.dbCache.set(gid, group);
-      this.dbDirty[gid] = true;
       return group[id];
     }
   });
@@ -74,10 +73,10 @@ export class ChunkedStore extends IStore {
       this.dbDirty[gid] = true;
       group[id] = entry;
     }
-    this.writeGroup(gid);
-    // if we batch A TON of set calls, ui can lag a lot, so we request a ui update every 1000 calls
+    if(!this.dbWritingPaused) this.writeGroup.callLater(gid);
+    // if we batch A TON of set calls, ui can lag a lot, so we request a ui update every 2000 calls
     this.entryWriteCounter++;
-    if(this.entryWriteCounter > 1000) {
+    if(this.entryWriteCounter > 2000) {
       this.entryWriteCounter = 0;
       await delay(0);
     }