+++ changelog.md
@@ -3,6 +3,8 @@ Fixes some minor bugs in some of 4.2
 
 ## Engine
 - Bug: Changelog was not exported.
+- Bug: Did not wait for the service worker to initialize before using it.
+- Bug: Queue error causing ChunkedStores to fail.
 
 ## Elemental 4.2.0 (2020-12-04)
 Fixes some major bugs and adding the first version of the recipe trees! Big thanks to Zelo(@Zolo101) for helping out.